backend/requirements.txt
# API & server
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9

# Mistral
mistralai==1.2.5

# Medical imaging model
torch==2.4.0
transformers==4.46.0
accelerate==0.34.0
Pillow==10.4.0

# PDF handling
pdfplumber==0.11.4
pdf2image==1.17.0

# Vector DB & embeddings
chromadb==0.5.15

# Utilities
python-dotenv==1.0.1
httpx==0.27.2
pydantic==2.9.2
pydantic-settings==2.5.2
